refactor additional backend classes for livy serializer compatibility
